Move the build-icons package in the packages folder.
