+++ packages/build-icons/src/index.mjs
@@ -1,0 +1,3 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as base64SVG } from './utils/base64SVG.mjs'
+export { default as getIconMetaData } from './utils/getIconMetaData.mjs'
