+++ packages/build-icons/src/render/default-attrs.json
