+++ packages/build-icons/src/utils/base64SVG.mjs
