--- packages/icons/internal/build-icons/index.mjs
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as getIconMetaData } from './utils/getIconMetaData.mjs';
