+++ packages/icons/package.json
@@ -1,7 +1,11 @@
 {
   "name": "icons",
-  "dependencies": {
-    "typescript": "^5.2.2"
+  "main": "./src/icons/index.ts",
+  "scripts": {
+    "build:icons": "build-icons --templateSrc=./scripts/exportTemplate.mjs --renderUniqueKey --iconFileExtension=.ts --exportFileName=index.ts"
   },
-  "main": "./src/icons/index.ts"
+  "dependencies": {
+    "typescript": "^5.2.2",
+    "@supabase/build-icons": "*"
+  }
 }
